feat(bot): enhance player search and rival display features
- Updated PlayerSearchResult to include first name, last name, and ratings for singles and doubles.
- Improved rival display logic to include player full name and added a filter for all disciplines.
- Introduced utility classes for generating player profile links and tournament result labels.
- Enhanced the UpdateDispatcher to provide clearer user feedback and improved error messages.
- Added unit tests for new functionalities and updated existing tests to reflect changes in data structures.

+++ core/src/main/java/ru/badmintonlab/core/repository/projection/RivalView.java
@@ -15,7 +15,7 @@ public interface RivalView {
 
     String getCity();
 
-    Short getWins();
+    Integer getWins();
 
-    Short getLosses();
+    Integer getLosses();
 }
